net-misc/dhcp: pull upstream ebuild 4.2.2-r1, take two
This version is only used in certain networking tests, and not in
production images. It is needed for linux-3.0 as it has rewritten
build system which doesn't break on linux3.

It needs to be explicitly unmasked by package.unmask, because our
version of portage snapshot hard masks it.

Compared to upstream, this ebuild is x86, arm, amd64 keyworded.

BUG=chromium-os:18442
TEST=build it on host and board

Change-Id: Ieb4914ae9274be397b1cefa90857eed24c14912f
Reviewed-on: http://gerrit.chromium.org/gerrit/6813
Tested-by: Zdenek Behan <zbehan@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>