libtool.eclass: import latest from upstream
Libtool has a bug where it assumes the default lib path is always
"/lib /usr/lib", and it searches /etc/ld.so.conf for additional
paths.  This does not play well when cross-compiling for a multilib
which uses a path other than /lib (like x86_64 which uses /lib64).
So libtool ends up adding -Wl,-rpath,/usr/lib64 when linking libs
which means it searches the chroot which means it can easily lead
to link failure.

So import the latest libtool.eclass and ELT-patches from upstream
Gentoo which has added a patch specifically to address this: the
sys-lib-dlsearch patch.  We have to import all the other files as
libtool.eclass only searches the ELT-patches/ subdir from the tree
from which it is loaded.

BUG=chromium-os:20636
TEST=`emerge-amd64-generic app-crypt/tpm-tools` now works
TEST=`./build_packages && ./build_image` for x86-alex still works

Change-Id: Id10c732cfd4f13df995f32af950a341e8c7bbc6c
Reviewed-on: http://gerrit.chromium.org/gerrit/8648
Reviewed-by: David James <davidjames@chromium.org>
Commit-Ready: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>