xmlrpc-c: import latest version from upstream Gentoo
The upstream ebuild added a build time patch so it works with newer
versions of curl (due to curl/types.h being removed).  Otherwise,
this is the same ebuild as we already have in the portage/ overlay.

This fixes the chromiumos-sdk buildbot which builds everything from
source and so is the only thing to notice this error so far.

BUG=None
TEST=`emerge xmlrpc-c` now works

Change-Id: I5e1b2a05b9491435b1fafe9bf538ef1a6f4228b7
Reviewed-on: https://gerrit.chromium.org/gerrit/12919
Reviewed-by: David James <davidjames@chromium.org>
Commit-Ready: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>