cmake: upgraded the cmake Portage package
The CL upgrades dev-util/cmake to version 2.8.6-r4 on amd64. The
update is necessary because I need to compile a library (OpenCV)
for a factory test. Besides that, the current version (2.6.4) is
more than 2 years old and therefore worthwhile for an upgrade.

There was an unsuccessful attempt to upgrade to 2.8.4-r1. This CL
tries to update to 2.8.6-r4 instead since it is the newest cmake
version in Gentoo.

Another corresponding CL on 'package.provided' will be carried
out before this one so the cmake won't be dragged into the build
list on platforms again.

BUG=chromium-os:23747
TEST='emerge cmake' still works
TEST='emerge poppler' still works
TEST='emerge-x86-alex libdivsufsort' still works
CQ-DEPEND=Ie8f1d4fa19586dbc7df150cbe9a970ecb43bc1d5

Change-Id: If9104c914a63033fcd47d3dd9863522e9e23ed12
Reviewed-on: https://gerrit.chromium.org/gerrit/13486
Reviewed-by: Rong Chang <rongchang@chromium.org>
Commit-Ready: Tai-Hsu Lin <sheckylin@chromium.org>
Tested-by: Tai-Hsu Lin <sheckylin@chromium.org>