Reland: sshfs-fuse: pull in version 2.4 from upstream
The upstream version 2.4 adds `-o slave`, which allows us to spawn
sshfs attached to the server end of an SSH pipe. As a result files
on your workstation can easily be pushed to a chromium-os device,
rather than needing to open a terminal on the device and SSH back.

This relands a (slightly modified) version of:
https://gerrit.chromium.org/gerrit/17724

BUG=chromium-os:25146
TEST=clean build and install test image on x86-alex
TEST=ran chrome over sshfs with and without -o slave
TEST=cbuildbot arm-tegra2-full

Change-Id: I6318a373067ed5166700d3c362ba16b34ad87797
Reviewed-on: https://gerrit.chromium.org/gerrit/17829
Commit-Ready: Chris Wolfe <cwolfe@chromium.org>
Reviewed-by: Chris Wolfe <cwolfe@chromium.org>
Tested-by: Chris Wolfe <cwolfe@chromium.org>