parted: Import parted-3.1 from upstream.
This CL imports sys-block/parted-3.1 from upstream, which will not be
used until all packages that depend on sys-apps/parted are switched to
depend on sys-block/parted.

parted-3.0 was previously imported, but due to a crash in partprobe,
parted-3.1 is imported to resolve the issue.

BUG=chromium-os:30350
TEST=Tested the following:
1. `emerge-{x86,amd64,arm}-generic parted` still builds
   sys-apps/parted-1.9.0-r1
2. `emerge-{x86,amd64,arm}-generic sys-block/parted` builds
   sys-block/parted-3.1 successfully after
   `emerge --unmerge sys-apps/parted`

Change-Id: I41ffe6a08184a3d3a1c9c64c96da57077120a08d
Reviewed-on: https://gerrit.chromium.org/gerrit/21705
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Ready: Ben Chan <benchan@chromium.org>
Tested-by: Ben Chan <benchan@chromium.org>