Store uprev x11-drivers/xf86-input-synaptics in portage-stable
Chromium OS images use xf86-input-cmt for the multitouch trackpads in
official Chromebooks.

For the trackpads on non-chromebooks, the X.org maintained
xf86-input-synaptics driver is also included in the chromium os image.

Currently, this uses a (stock) out-of-date pre-release ebuild maintanied
in the chromiumos-overlay.

It is more appropriate to store unmodified ebuilds in the portage-stable
overlay.

As a bonus, uprev to latest stable[*].

[*] This patch also marks the upstream version as stable for arm.

A separate commit removes the xf86-input-synaptics ebuild from the
chromiumos-overlay.

BUG=chromium-os:31764
TEST=cbuildbot --remote -g <gerrit ID> {x86,amd64,arm}-generic-full

Change-Id: I6e8fd3d6c35f06c677b8839700e0c7bdb224201f
Reviewed-on: https://gerrit.chromium.org/gerrit/33759
Commit-Ready: Daniel Kurtz <djkurtz@chromium.org>
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>