bump(sys-apps/net-tools): sync with upstream
We're running with a rather old copy of the ebuild.
While we might replace net-tools at some point, in the meanwhile we
should probably keep it reasonably up to date.

Note: this does drop the `ypdomainname` and `nisdomainname` binaries
from `/usr/bin` (since they are now behind the default-off nis use flag)