dev-vcs/git: update package for CVEs
See https://marc.info/?l=git&m=157600115215285&w=2

This omits the 2.24+ and 9999+ ebuilds, they use EAPI 7 and our perl
eclass does not. These could be updated in the future if needed, but
it's not worth the time now. Those ebuild are not marked stable anyway,
so the image/sdk contents are not affected by their absense.